feat: hide multi-team UI and pin to a seeded default team
Multi-team support remains in the DB, server routes, and HTTP APIs; only
the UI is single-team. A default team (id 00000000-0000-0000-0000-000000000001,
slug 'default') is seeded at server boot from the Startup template, and
the bootstrap admin is auto-added as a board member.

- shared: add DEFAULT_TEAM_ID/SLUG/NAME/TEMPLATE_NAME constants
- server: extract createTeam service from the POST /api/teams handler so
  the route and the boot-time seeder share one code path
- server: seedDefaultTeam runs after services wire up; idempotent
- web: drop team rail, /teams/new, and the global /inbox route; redirect
  / and /teams to /teams/default
- web: relocate global Settings link and ThemeSwitcher into the team
  sidebar footer
- web: trim now-unused team hooks
- tests: remove team-crud e2e specs; update route-redirects expectation
