+++ packages/server/src/routes/auth.ts
@@ -1,3 +1,4 @@
+import { DEFAULT_TEAM_ID, MemberType } from '@hezo/shared';
 import { Hono } from 'hono';
 import { err, ok } from '../lib/response';
 import type { Env } from '../lib/types';
@@ -32,8 +33,33 @@ authRoutes.post('/auth/token', async (c) => {
 			"INSERT INTO users (display_name, is_superuser) VALUES ('Admin', true) RETURNING id",
 		);
 		userId = inserted.rows[0].id;
+		await addUserToDefaultTeam(db, userId);
 	}
 
 	const token = await signBoardJwt(masterKeyManager, userId);
 	return ok(c, { token }, 200);
 });
+
+async function addUserToDefaultTeam(
+	db: import('@electric-sql/pglite').PGlite,
+	userId: string,
+): Promise<void> {
+	const existing = await db.query(
+		`SELECT m.id FROM members m
+		 JOIN member_users mu ON mu.id = m.id
+		 WHERE mu.user_id = $1 AND m.team_id = $2`,
+		[userId, DEFAULT_TEAM_ID],
+	);
+	if (existing.rows.length > 0) return;
+
+	const member = await db.query<{ id: string }>(
+		`INSERT INTO members (team_id, member_type, display_name)
+		 VALUES ($1, $2, (SELECT display_name FROM users WHERE id = $3))
+		 RETURNING id`,
+		[DEFAULT_TEAM_ID, MemberType.User, userId],
+	);
+	await db.query(`INSERT INTO member_users (id, user_id, role) VALUES ($1, $2, 'board')`, [
+		member.rows[0].id,
+		userId,
+	]);
+}
